benchmark sw serde
